backend/requirements.txt
django==5.0
djangorestframework==3.15.1
django-cors-headers==4.3.1